Add files via upload
A little joke you left as a foundation for this technical skills thing. You want people to do this - use this as a learning tool, yet you don't even bother with checking your files and their contents first? 
What a joke. Can't take this seriously anymore and with my condition to top it off.